Replace FNV-1a with memhash in Map (#14)
Includes the following changes:

* Change hash64 return type

It makes no sense to trim the returned result to uint32
in hash64. Moreover, that may make hash code distribution
worse.

* Replace FNV-1a hash function with memhash

Replaces FNV-1a 32-bit function with faster built-in 64-bit
memhash function which is internally used in Golang standard
library.

+++ counter.go
@@ -53,7 +53,7 @@ func (c *Counter) Add(delta int64) {
 	if !ok {
 		t = new(ptoken)
 		// Since cstripes is a power of two, we can use & instead of %.
-		t.idx = hash64(uintptr(unsafe.Pointer(t))) & (cstripes - 1)
+		t.idx = uint32(hash64(uintptr(unsafe.Pointer(t))) & (cstripes - 1))
 	}
 	stripe := &c.stripes[t.idx]
 	atomic.AddInt64(&stripe.c, delta)
